requirements.txt
aldryn-apphooks-config==0.6.0
apturl==0.5.2
asgiref==3.2.7
asn1crypto==0.24.0
blis==0.2.4
Brlapi==0.6.6
certifi==2018.1.18
chardet==3.0.4
ChatterBot==1.0.5
chatterbot-corpus==1.2.0
click==7.1.2
command-not-found==0.3
cryptography==2.1.4
cupshelpers==1.0
cymem==2.0.3
defer==1.0.6
distro-info===0.18ubuntu0.18.04.1
dj-database-url==0.5.0
Django==2.2.12
django-appdata==0.3.0
django-classy-tags==1.0.0
django-cms==3.7.2
django-filer==1.7.1
django-formtools==2.2
django-js-asset==1.2.2
django-meta==1.6.1
django-mptt==0.11.0
django-parler==2.0.1
django-polymorphic==2.1.2
django-sekizai==1.1.0
django-sortedm2m==3.0.0
django-taggit==1.2.0
django-taggit-autosuggest==0.3.8
django-taggit-templatetags==0.2.5
django-templatetag-sugar==1.0
django-treebeard==4.3.1
djangocms-admin-style==1.5.0
djangocms-apphook-setup==0.4.1
djangocms-attributes-field==1.2.0
djangocms-blog==1.1.1
djangocms-bootstrap4==1.6.0
djangocms-file==2.4.0
djangocms-googlemap==1.4.0
djangocms-icon==1.5.0
djangocms-installer==1.2.3
djangocms-link==2.6.1
djangocms-picture==2.4.0
djangocms-snippet==2.3.0
djangocms-style==2.3.0
djangocms-text-ckeditor==3.9.0
djangocms-video==2.3.0
dlib==19.19.0
easy-thumbnails==2.7
face-recognition==1.3.0
face-recognition-models==0.3.0
html5lib==1.0.1
httplib2==0.9.2
idna==2.6
joblib==0.15.1
keyring==10.6.0
keyrings.alt==3.0
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
louis==3.5.0
lxml==4.5.0
macaroonbakery==1.1.3
Mako==1.0.7
MarkupSafe==1.0
mathparse==0.1.2
murmurhash==1.0.2
netifaces==0.10.4
nltk==3.5
numpy==1.18.4
oauth==1.0.1
olefile==0.45.1
pexpect==4.2.1
Pillow==5.1.0
Pint==0.11
plac==0.9.6
preshed==2.0.1
protobuf==3.0.0
pycairo==1.16.2
pycrypto==2.6.1
pycups==1.9.73
pygobject==3.26.1
pymacaroons==0.13.0
pymongo==3.10.1
PyNaCl==1.1.2
pyRFC3339==1.0
python-apt==1.6.5+ubuntu0.2
python-dateutil==2.7.5
python-debian==0.1.32
pytz==2020.1
pyxdg==0.25
PyYAML==3.13
regex==2020.5.14
reportlab==3.4.0
requests==2.18.4
requests-unixsocket==0.1.5
screen-resolution-extra==0.0.0
SecretStorage==2.3.1
simplejson==3.13.2
six==1.14.0
spacy==2.1.9
SQLAlchemy==1.2.19
sqlparse==0.3.1
srsly==1.0.2
system-service==0.3
systemd-python==234
thinc==7.0.8
tqdm==4.46.0
tzlocal==2.1
ubuntu-drivers-common==0.0.0
ufw==0.35
unattended-upgrades==0.1
Unidecode==1.1.1
urllib3==1.22
usb-creator==0.3.3
wadllib==1.3.2
wasabi==0.6.0
webencodings==0.5.1
xkit==0.0.0
zope.interface==4.3.2
